+++ cmake_files/build_options.cmake
@@ -18,6 +18,7 @@ option(USE_XRANDR "Enable Xrandr support"              ON )
 # Temp and hidden-ish, to make it easier to deorbit all at once
 option(USE_CAPTIVE "Enable captive CTWM support" OFF )
 option(USE_VSCREEN "Enable VirtualScreens support" OFF )
+option(USE_WINBOX  "Enable WindowBox support" OFF )
 
 
 
@@ -175,3 +176,12 @@ if(USE_VSCREEN)
 	message(WARNING "VirtualScreens will not be supported in future versions")
 else()
 endif(USE_VSCREEN)
+
+
+# WindowBox's
+if(USE_WINBOX)
+	message(WARNING "WindowBox will not be supported in future versions")
+
+	list(APPEND CTWMSRC windowbox.c)
+else()
+endif(USE_WINBOX)