+++ functions_internal.h
@@ -167,7 +167,9 @@ DFHANDLER(quit);
 DFHANDLER(restart);
 DFHANDLER(beep);
 DFHANDLER(trace);
+#ifdef WINBOX
 DFHANDLER(fittocontent);
+#endif
 DFHANDLER(showbackground);
 DFHANDLER(raiseicons);
 DFHANDLER(rescuewindows);